added file uniqueness options
Files that are confirmed conflicts (same name but different content)
are moved with deconflicted names. In essence, this means a unique
string is added to the file to make the name unique when performing the
move to the destination dir. Now you can have the uniques string added
in one of 3 places.

Fixed some newline handling.

Fixed a bug regarding the handling of duplicate files.